Follow fog-core best practices
Following https://github.com/fog/fog-core/pull/229, this gem needs to be namespaced to avoid deprecation warnings. This commit moves the files around to fit the better namespacing techniques. It additionally removes the `.ruby-version` file (so as to allow more easily testing and developing other Ruby versions) and adds Ruby 2.5.1 to the travis test suite.